Safari-609.1.11.1/LayoutTests/editing/pasteboard/data-transfer-set-data-sanitize-url-when-copying-in-null-origin-expected.txt

This tests calling setData with a malformed URL in a null origin document. The malformed value should not be readable in another document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS urlReadInSameDocument is "http://webkit.org/b/🤔?x=8 + 6<"
PASS htmlReadInSameDocument is "testing"
PASS JSON.stringify(typesInSameDocument) is "[\"text/uri-list\",\"text/html\"]"
PASS JSON.stringify(itemsInSameDocument) is "[{\"kind\":\"string\",\"type\":\"text/uri-list\"},{\"kind\":\"string\",\"type\":\"text/html\"}]"
PASS event.clipboardData.getData("url") is "http://webkit.org/b/%F0%9F%A4%94?x=8%20+%206%3C"
PASS JSON.stringify(event.clipboardData.types) is "[\"text/uri-list\",\"text/html\"]"
PASS JSON.stringify(Array.from(event.clipboardData.items).map((item) => ({kind: item.kind, type: item.type}))) is "[{\"kind\":\"string\",\"type\":\"text/uri-list\"},{\"kind\":\"string\",\"type\":\"text/html\"}]"
PASS successfullyParsed is true

TEST COMPLETE
